+++ src/MyNeuralNetwork.Domain/Entities/Support/MyRandom.cs
@@ -1,21 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MyNeuralNetwork.Domain.Entities.Support
 {
     public static class MyRandom
     {
 
-        public static float Range(float v, float v1)
+        public static double Range(double v, double v1)
         {
             var _random = new Random();
-            return ApplyFormula(v, v1, (float)_random.NextDouble());
+            return ApplyFormula(v, v1, (double)_random.NextDouble());
         }
 
-        private static float ApplyFormula(float v1, float v2, float rand)
+        private static double ApplyFormula(double v1, double v2, double rand)
         {
             return (v2 - v1) * rand + v1;
         }
